Create Readme.txt
Nowadays, people use their smartphones to collect their data relating to the number of steps taken, number of calories consumed, the kind of meals taken, amount of water taken, etc. This project involves deriving the insights from population survey and using these insights to detect the human activity being performed. For example, a person’s walking, running, sitting, etc. is recognized by using the collected data through smartphone. This can help as an assistive technology in healthcare sectors and eldercare when ensemble with other technologies like Internet of Things (IOT). In this paper, we discuss various machine learning algorithms to derive different kinds of information from the given dataset and compare the performances of the each of the algorithms.